Fix Docker (#6)
* Simplify the installation of system dependencies in Dockerfile

The system dependencies installation in Dockerfile was updated to use the `apk add` command directly, eliminating the need of using the virtual temporary build dependencies. This makes the docker build process cleaner and simpler.

* Update GitHub Actions to run on pull requests

The GitHub Actions workflow has been updated to also run on pull requests targeting the main branch. Additionally, the configuration has been modified so that Docker images are only pushed on commits to the main branch.

---------

Co-authored-by: Jag_k <jag-k@users.noreply.github.com>